Correcting var names and types

+++ PudimDouroAPI/src/main/java/br/com/mountainfortress/PudimDouroAPI/repository/RegistrationTokenRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface RegistrationTokenRepository extends JpaRepository<RegistrationToken, Long> {
+public interface RegistrationTokenRepository extends JpaRepository<RegistrationToken, Integer> {
     @Query(value = "FROM RegistrationToken rt WHERE rt.active = TRUE")
     List<RegistrationToken> findActiveToken();
 }
